migrate useK8sModel(s) to SDK
