+++ frontend/packages/console-dynamic-plugin-sdk/src/api/core-api.ts
@@ -1,7 +1,5 @@
 /* eslint-disable */
-import {
-  UseActivePerspective,
-} from '../extensions/console-types';
+import { UseActivePerspective } from '../extensions/console-types';
 
 export const useActivePerspective: UseActivePerspective = require('@console/dynamic-plugin-sdk/src/perspective/useActivePerspective')
   .default;
