+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/k8s/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useK8sModel } from './useK8sModel';
+export { useK8sModels } from './useK8sModels';
